requirements.txt
# Core data science libraries
pandas>=1.5.0
numpy>=1.21.0
scipy>=1.9.0

# Data collection and APIs
requests>=2.28.0
beautifulsoup4>=4.11.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0

# Machine Learning
scikit-learn>=1.2.0
tensorflow>=2.12.0

# Time Series Analysis
statsmodels>=0.14.0
prophet>=1.1.4

# MLOps and Model Management
mlflow>=2.5.0
joblib>=1.2.0

# Configuration and utilities
pyyaml>=6.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.65.0

# API development
fastapi>=0.100.0
uvicorn[standard]>=0.22.0

# Dashboard
streamlit>=1.25.0

# Database support
sqlalchemy>=2.0.0

# Logging and monitoring
prometheus-client>=0.17.0